+++ sdk/js/server/src/wasm-loader.ts
@@ -1,9 +1,0 @@
-let wasmModule: any;
-
-export async function loadActraWasm() {
-    if (!wasmModule) {
-        wasmModule = await import("../pkg/server/actra_wasm.js");
-    }
-
-    return wasmModule;
-}